@@ -1,10 +1,9 @@
-from ._version import __version__
-
 from ._bmi import (
 {%- for babelized_class, _ in cookiecutter.components|dictsort %}
     {{ babelized_class }},
 {%- endfor %}
 )
+from ._version import __version__
 
 __all__ = [
     "__version__",
